feature: adapted db into celery implementation.
