+++ docker-compose.yaml
@@ -2,7 +2,8 @@ version: "3.2"
 services:
   celery:
     build:
-      context: './batch'
+      context: ./
+      dockerfile: ./batch/Dockerfile
     volumes:
       - ./shared:/shared/
     depends_on:
@@ -42,6 +43,8 @@ services:
       - POSTGRES_PASSWORD=1234
       - POSTGRES_DB=postgres
       - PGDATA=/var/lib/postgresql/data/ifpvdl/
+    networks:
+      network:
   nginx:
     image: nginx:latest
     ports:
